backend/pyproject.toml
[project]
name = "resume-maker-backend"
version = "0.1.0"
description = "AI-powered resume customization service backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community>=0.4.0",
    "openai>=1.54.0",
    "PyPDF2>=3.0.1",
    "pypdf>=5.0.0",
    "reportlab>=4.2.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
